backend/requirements.txt
aiosmtplib==4.0.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-oauth==0.15.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.5.0
makefun==1.15.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.12
passlib==1.7.4
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.1
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spark-parser==1.8.9
SQLAlchemy==2.0.36
starlette==0.41.3
stripe==11.3.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uncompyle6==3.9.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
xdis==6.1.2
